feat(blog2.0):bug

+++ src/main/java/com/huorong/Controller/blog2/ShareController.java
@@ -24,10 +24,10 @@ public class ShareController {
     @RequestMapping("toShare")
     public Result regist(@RequestParam Map params) {
         String shareUrl = MapUtils.getStr(params, "shareUrl");
-        if (StringUtils.isEmpty(shareUrl)) {
-            Result.build("1", "error");
+        String userId = MapUtils.getStr(params, "userId");
+        if (StringUtils.isEmpty(shareUrl) || StringUtils.isEmpty(userId)) {
+            return Result.build("1", "error");
         }
-
         return shareService.insertShare(params) ? Result.build("0", "ok") : Result.build("1", "error");
     }
 
@@ -38,6 +38,10 @@ public class ShareController {
 
     @RequestMapping("uploadImag")
     public Result uploadImag(@RequestParam Map params) {
+        String userId = MapUtils.getStr(params, "userId");
+        if (StringUtils.isEmpty(userId)) {
+            return Result.build("1", "error");
+        }
         return Result.build("0", "ok", MapUtils.of("image", shareService.uploadImage(params)));
     }
 
@@ -49,6 +53,9 @@ public class ShareController {
 
     @RequestMapping("delteShares")
     public Result deleteMyShare(@RequestParam String userId, @RequestParam String shareId) {
+        if (StringUtils.isEmpty(userId)) {
+            return Result.build("1", "error");
+        }
         return shareService.deleteMyShare(userId, shareId) ? Result.build("0", "ok") : Result.build("1", "error");
     }
